Move qglue data/link parsing to glue.core
