+++ glue/core/parsers/__init__.py
@@ -1,0 +1,1 @@
+from .parsers import parse_data, parse_links  # noqa
